Fix 5 critical data integrity issues and consolidate service interfaces
CRITICAL FIXES:
1. Enable SQLite foreign key enforcement (PRAGMA foreign_keys = ON)
2. Add transactional integrity to billing operations (GenerateInvoice, AddPayment)
3. Fix floating-point precision loss in panel price distribution (decimal not double)
4. Prevent payment race condition and add overpayment guard
5. Make result saving atomic with transaction + validate TypeId belongs to order

REFACTORING:
- Move IBillingService and other service interfaces into Core/Interfaces/IServices.cs
- Add new unit tests (DashboardViewModel, ReferenceRangeEvaluator)
- Consolidate DatabaseInitializer and seed logic
- Extend V1__init.sql schema for expanded clinical data
- Improve SecureConfigurationManager with DPAPI error handling

+++ LabSystem.Core/Interfaces/ITestOrderRepository.cs
@@ -10,6 +10,6 @@ namespace LabSystem.Core.Interfaces
         Task<IEnumerable<TestOrder>> GetOrdersForPatientAsync(int patientId, CancellationToken cancellationToken = default(CancellationToken));
         Task<IEnumerable<TestOrder>> GetByStatusAsync(string status, CancellationToken cancellationToken = default(CancellationToken));
         Task AddOrderWithTestTypesAsync(TestOrder order, List<int> testTypeIds, CancellationToken cancellationToken = default(CancellationToken));
-        Task UpdateOrderTestPricingAsync(int orderId, int typeId, int? packageId, double billedCost, CancellationToken cancellationToken = default(CancellationToken));
+        Task UpdateOrderTestPricingAsync(int orderId, int typeId, int? packageId, decimal billedCost, CancellationToken cancellationToken = default(CancellationToken));
     }
 }
